Reworked sword attacks and player can attack more than 1 enemy at a time
The player can now attack any enemy in a group if they are close enough and within range of the weapon. This required a rework of how attacks are handled, the attack system will no longer require that the player hovers over an enemy, but, instead the weapon has a box collider.