Gold coins move towards player, player character screen
Started on a new GUI element that displays the player's stats, along with the amount of money they have. Also made it so that gold coins do not get added to the inventory, and will instead increment a value. Furthermore, they also move towards the player when they are close enough.